third_party/WebKit/LayoutTests/animations/animations-parsing-expected.txt

Test the parsing and the computed style values of the animations properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Valid animation-name values.
PASS computedStyle.animationName is 'none'
PASS computedStyle.webkitAnimationName is 'none'
PASS Object.keys(style).indexOf('animationName') is not -1
PASS Object.keys(style).indexOf('webkitAnimationName') is -1
PASS style.animationName is 'none'
PASS computedStyle.animationName is 'none'
PASS style.webkitAnimationName is 'none'
PASS computedStyle.webkitAnimationName is 'none'
PASS style.animationName is 'none, none'
PASS computedStyle.animationName is 'none, none'
PASS style.webkitAnimationName is 'none, none'
PASS computedStyle.webkitAnimationName is 'none, none'
PASS style.animationName is 'myAnim'
PASS computedStyle.animationName is 'myAnim'
PASS style.webkitAnimationName is 'myAnim'
PASS computedStyle.webkitAnimationName is 'myAnim'
PASS style.animationName is 'background-position, font-size'
PASS computedStyle.animationName is 'background-position, font-size'
PASS style.webkitAnimationName is 'background-position, font-size'
PASS computedStyle.webkitAnimationName is 'background-position, font-size'
PASS style.animationName is 'background-position, font-size, color'
PASS computedStyle.animationName is 'background-position, font-size, color'
PASS style.webkitAnimationName is 'background-position, font-size, color'
PASS computedStyle.webkitAnimationName is 'background-position, font-size, color'
PASS style.animationName is 'all, font-size, color'
PASS computedStyle.animationName is 'all, font-size, color'
PASS style.webkitAnimationName is 'all, font-size, color'
PASS computedStyle.webkitAnimationName is 'all, font-size, color'
PASS style.animationName is 'inherit'
PASS computedStyle.animationName is 'none'
PASS style.webkitAnimationName is 'inherit'
PASS computedStyle.webkitAnimationName is 'none'
PASS style.animationName is 'initial'
PASS computedStyle.animationName is 'none'
PASS style.webkitAnimationName is 'initial'
PASS computedStyle.webkitAnimationName is 'none'
Invalid animation-name values.
PASS style.animationName is ''
PASS computedStyle.animationName is 'none'
PASS style.webkitAnimationName is ''
PASS computedStyle.webkitAnimationName is 'none'
PASS style.animationName is ''
PASS computedStyle.animationName is 'none'
PASS style.webkitAnimationName is ''
PASS computedStyle.webkitAnimationName is 'none'
PASS style.animationName is ''
PASS computedStyle.animationName is 'none'
PASS style.webkitAnimationName is ''
PASS computedStyle.webkitAnimationName is 'none'
Valid animation-duration values.
PASS computedStyle.animationDuration is '0s'
PASS computedStyle.webkitAnimationDuration is '0s'
PASS Object.keys(style).indexOf('animationDuration') is not -1
PASS Object.keys(style).indexOf('webkitAnimationDuration') is -1
PASS style.animationDuration is '0s'
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is '0s'
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is '5s'
PASS computedStyle.animationDuration is '5s'
PASS style.webkitAnimationDuration is '5s'
PASS computedStyle.webkitAnimationDuration is '5s'
PASS style.animationDuration is '10ms'
PASS computedStyle.animationDuration is '0.01s'
PASS style.webkitAnimationDuration is '10ms'
PASS computedStyle.webkitAnimationDuration is '0.01s'
PASS style.animationDuration is '10ms, 20s'
PASS computedStyle.animationDuration is '0.01s, 20s'
PASS style.webkitAnimationDuration is '10ms, 20s'
PASS computedStyle.webkitAnimationDuration is '0.01s, 20s'
Invalid animation-duration values.
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
PASS style.animationDuration is ''
PASS computedStyle.animationDuration is '0s'
PASS style.webkitAnimationDuration is ''
PASS computedStyle.webkitAnimationDuration is '0s'
Valid animation-timing-function values.
PASS computedStyle.animationTimingFunction is 'ease'
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS Object.keys(style).indexOf('animationTimingFunction') is not -1
PASS Object.keys(style).indexOf('webkitAnimationTimingFunction') is -1
PASS style.animationTimingFunction is 'linear'
PASS computedStyle.animationTimingFunction is 'linear'
PASS style.webkitAnimationTimingFunction is 'linear'
PASS computedStyle.webkitAnimationTimingFunction is 'linear'
PASS style.animationTimingFunction is 'ease'
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is 'ease'
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is 'ease-in'
PASS computedStyle.animationTimingFunction is 'ease-in'
PASS style.webkitAnimationTimingFunction is 'ease-in'
PASS computedStyle.webkitAnimationTimingFunction is 'ease-in'
PASS style.animationTimingFunction is 'ease-out'
PASS computedStyle.animationTimingFunction is 'ease-out'
PASS style.webkitAnimationTimingFunction is 'ease-out'
PASS computedStyle.webkitAnimationTimingFunction is 'ease-out'
PASS style.animationTimingFunction is 'ease-in-out'
PASS computedStyle.animationTimingFunction is 'ease-in-out'
PASS style.webkitAnimationTimingFunction is 'ease-in-out'
PASS computedStyle.webkitAnimationTimingFunction is 'ease-in-out'
PASS style.animationTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.25, 0.1, 0.25, 1)'
PASS style.animationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.animationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 0, 0.23, 3)'
PASS style.animationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 3)'
PASS style.animationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.1, 5, 0.23, 0)'
PASS style.animationTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.2, 2, 0.2, -0.4)'
PASS style.animationTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS computedStyle.animationTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS style.webkitAnimationTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS computedStyle.webkitAnimationTimingFunction is 'cubic-bezier(0.2, -2, 0.2, 0.4)'
PASS style.animationTimingFunction is 'step-start'
PASS computedStyle.animationTimingFunction is 'step-start'
PASS style.webkitAnimationTimingFunction is 'step-start'
PASS computedStyle.webkitAnimationTimingFunction is 'step-start'
PASS style.animationTimingFunction is 'step-end'
PASS computedStyle.animationTimingFunction is 'step-end'
PASS style.webkitAnimationTimingFunction is 'step-end'
PASS computedStyle.webkitAnimationTimingFunction is 'step-end'
PASS style.animationTimingFunction is 'steps(3, end)'
PASS computedStyle.animationTimingFunction is 'steps(3, end)'
PASS style.webkitAnimationTimingFunction is 'steps(3, end)'
PASS computedStyle.webkitAnimationTimingFunction is 'steps(3, end)'
PASS style.animationTimingFunction is 'steps(4, end)'
PASS computedStyle.animationTimingFunction is 'steps(4, end)'
PASS style.webkitAnimationTimingFunction is 'steps(4, end)'
PASS computedStyle.webkitAnimationTimingFunction is 'steps(4, end)'
PASS style.animationTimingFunction is 'steps(5, start)'
PASS computedStyle.animationTimingFunction is 'steps(5, start)'
PASS style.webkitAnimationTimingFunction is 'steps(5, start)'
PASS computedStyle.webkitAnimationTimingFunction is 'steps(5, start)'
PASS style.animationTimingFunction is 'frames(5)'
PASS computedStyle.animationTimingFunction is 'frames(5)'
PASS style.webkitAnimationTimingFunction is 'frames(5)'
PASS computedStyle.webkitAnimationTimingFunction is 'frames(5)'
PASS style.animationTimingFunction is 'ease-in-out, ease-in'
PASS computedStyle.animationTimingFunction is 'ease-in-out, ease-in'
PASS style.webkitAnimationTimingFunction is 'ease-in-out, ease-in'
PASS computedStyle.webkitAnimationTimingFunction is 'ease-in-out, ease-in'
PASS style.animationTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.animationTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.webkitAnimationTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS computedStyle.webkitAnimationTimingFunction is 'ease-in-out, cubic-bezier(0.1, 0, 0.23, 0.4)'
PASS style.animationTimingFunction is 'steps(3, start), ease-in-out'
PASS computedStyle.animationTimingFunction is 'steps(3, start), ease-in-out'
PASS style.webkitAnimationTimingFunction is 'steps(3, start), ease-in-out'
PASS computedStyle.webkitAnimationTimingFunction is 'steps(3, start), ease-in-out'
Invalid animation-timing-function values.
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is ''
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is ''
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is 'step-middle'
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is 'step-middle'
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
PASS style.animationTimingFunction is 'steps(5, middle)'
PASS computedStyle.animationTimingFunction is 'ease'
PASS style.webkitAnimationTimingFunction is 'steps(5, middle)'
PASS computedStyle.webkitAnimationTimingFunction is 'ease'
Valid animation-delay values.
PASS computedStyle.animationDelay is '0s'
PASS computedStyle.webkitAnimationDelay is '0s'
PASS Object.keys(style).indexOf('animationDelay') is not -1
PASS Object.keys(style).indexOf('webkitAnimationDelay') is -1
PASS style.animationDelay is '0s'
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is '0s'
PASS computedStyle.webkitAnimationDelay is '0s'
PASS style.animationDelay is '5s'
PASS computedStyle.animationDelay is '5s'
PASS style.webkitAnimationDelay is '5s'
PASS computedStyle.webkitAnimationDelay is '5s'
PASS style.animationDelay is '10ms'
PASS computedStyle.animationDelay is '0.01s'
PASS style.webkitAnimationDelay is '10ms'
PASS computedStyle.webkitAnimationDelay is '0.01s'
PASS style.animationDelay is '-10ms'
PASS computedStyle.animationDelay is '-0.01s'
PASS style.webkitAnimationDelay is '-10ms'
PASS computedStyle.webkitAnimationDelay is '-0.01s'
PASS style.animationDelay is '-10ms, 20s'
PASS computedStyle.animationDelay is '-0.01s, 20s'
PASS style.webkitAnimationDelay is '-10ms, 20s'
PASS computedStyle.webkitAnimationDelay is '-0.01s, 20s'
Invalid animation-delay values.
PASS style.animationDelay is ''
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is ''
PASS computedStyle.webkitAnimationDelay is '0s'
PASS style.animationDelay is ''
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is ''
PASS computedStyle.webkitAnimationDelay is '0s'
PASS style.animationDelay is ''
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is ''
PASS computedStyle.webkitAnimationDelay is '0s'
PASS style.animationDelay is ''
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is ''
PASS computedStyle.webkitAnimationDelay is '0s'
PASS style.animationDelay is ''
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is ''
PASS computedStyle.webkitAnimationDelay is '0s'
PASS style.animationDelay is ''
PASS computedStyle.animationDelay is '0s'
PASS style.webkitAnimationDelay is ''
PASS computedStyle.webkitAnimationDelay is '0s'
Valid animation-iteration-count values.
PASS computedStyle.animationIterationCount is '1'
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS Object.keys(style).indexOf('animationIterationCount') is not -1
PASS Object.keys(style).indexOf('webkitAnimationIterationCount') is -1
PASS style.animationIterationCount is '4'
PASS computedStyle.animationIterationCount is '4'
PASS style.webkitAnimationIterationCount is '4'
PASS computedStyle.webkitAnimationIterationCount is '4'
PASS style.animationIterationCount is '2.5'
PASS computedStyle.animationIterationCount is '2.5'
PASS style.webkitAnimationIterationCount is '2.5'
PASS computedStyle.webkitAnimationIterationCount is '2.5'
PASS style.animationIterationCount is 'infinite'
PASS computedStyle.animationIterationCount is 'infinite'
PASS style.webkitAnimationIterationCount is 'infinite'
PASS computedStyle.webkitAnimationIterationCount is 'infinite'
PASS style.animationIterationCount is 'infinite, 3'
PASS computedStyle.animationIterationCount is 'infinite, 3'
PASS style.webkitAnimationIterationCount is 'infinite, 3'
PASS computedStyle.webkitAnimationIterationCount is 'infinite, 3'
PASS style.animationIterationCount is '0'
PASS computedStyle.animationIterationCount is '0'
PASS style.webkitAnimationIterationCount is '0'
PASS computedStyle.webkitAnimationIterationCount is '0'
PASS style.animationIterationCount is 'infinite, infinite'
PASS computedStyle.animationIterationCount is 'infinite, infinite'
PASS style.webkitAnimationIterationCount is 'infinite, infinite'
PASS computedStyle.webkitAnimationIterationCount is 'infinite, infinite'
Invalid animation-iteration-count values.
PASS style.animationIterationCount is ''
PASS computedStyle.animationIterationCount is '1'
PASS style.webkitAnimationIterationCount is ''
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS style.animationIterationCount is ''
PASS computedStyle.animationIterationCount is '1'
PASS style.webkitAnimationIterationCount is ''
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS style.animationIterationCount is ''
PASS computedStyle.animationIterationCount is '1'
PASS style.webkitAnimationIterationCount is ''
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS style.animationIterationCount is ''
PASS computedStyle.animationIterationCount is '1'
PASS style.webkitAnimationIterationCount is ''
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS style.animationIterationCount is ''
PASS computedStyle.animationIterationCount is '1'
PASS style.webkitAnimationIterationCount is ''
PASS computedStyle.webkitAnimationIterationCount is '1'
PASS style.animationIterationCount is ''
PASS computedStyle.animationIterationCount is '1'
PASS style.webkitAnimationIterationCount is ''
PASS computedStyle.webkitAnimationIterationCount is '1'
Valid animation-direction values.
PASS computedStyle.animationDirection is 'normal'
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS Object.keys(style).indexOf('animationDirection') is not -1
PASS Object.keys(style).indexOf('webkitAnimationDirection') is -1
PASS style.animationDirection is 'reverse'
PASS computedStyle.animationDirection is 'reverse'
PASS style.webkitAnimationDirection is 'reverse'
PASS computedStyle.webkitAnimationDirection is 'reverse'
PASS style.animationDirection is 'normal'
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is 'normal'
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is 'normal, alternate'
PASS computedStyle.animationDirection is 'normal, alternate'
PASS style.webkitAnimationDirection is 'normal, alternate'
PASS computedStyle.webkitAnimationDirection is 'normal, alternate'
PASS style.animationDirection is 'alternate'
PASS computedStyle.animationDirection is 'alternate'
PASS style.webkitAnimationDirection is 'alternate'
PASS computedStyle.webkitAnimationDirection is 'alternate'
PASS style.animationDirection is 'alternate-reverse'
PASS computedStyle.animationDirection is 'alternate-reverse'
PASS style.webkitAnimationDirection is 'alternate-reverse'
PASS computedStyle.webkitAnimationDirection is 'alternate-reverse'
Invalid animation-direction values.
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
PASS style.animationDirection is ''
PASS computedStyle.animationDirection is 'normal'
PASS style.webkitAnimationDirection is ''
PASS computedStyle.webkitAnimationDirection is 'normal'
Valid animation-play-state values.
PASS computedStyle.animationPlayState is 'running'
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS Object.keys(style).indexOf('animationPlayState') is not -1
PASS Object.keys(style).indexOf('webkitAnimationPlayState') is -1
PASS style.animationPlayState is 'paused'
PASS computedStyle.animationPlayState is 'paused'
PASS style.webkitAnimationPlayState is 'paused'
PASS computedStyle.webkitAnimationPlayState is 'paused'
PASS style.animationPlayState is 'running'
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is 'running'
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is 'running, running'
PASS computedStyle.animationPlayState is 'running, running'
PASS style.webkitAnimationPlayState is 'running, running'
PASS computedStyle.webkitAnimationPlayState is 'running, running'
PASS style.animationPlayState is 'paused, paused'
PASS computedStyle.animationPlayState is 'paused, paused'
PASS style.webkitAnimationPlayState is 'paused, paused'
PASS computedStyle.webkitAnimationPlayState is 'paused, paused'
PASS style.animationPlayState is 'paused, running'
PASS computedStyle.animationPlayState is 'paused, running'
PASS style.webkitAnimationPlayState is 'paused, running'
PASS computedStyle.webkitAnimationPlayState is 'paused, running'
Invalid animation-play-state values.
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
PASS style.animationPlayState is ''
PASS computedStyle.animationPlayState is 'running'
PASS style.webkitAnimationPlayState is ''
PASS computedStyle.webkitAnimationPlayState is 'running'
Valid animation-fill-mode values.
PASS computedStyle.animationFillMode is 'none'
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS Object.keys(style).indexOf('animationFillMode') is not -1
PASS Object.keys(style).indexOf('webkitAnimationFillMode') is -1
PASS style.animationFillMode is 'forwards'
PASS computedStyle.animationFillMode is 'forwards'
PASS style.webkitAnimationFillMode is 'forwards'
PASS computedStyle.webkitAnimationFillMode is 'forwards'
PASS style.animationFillMode is 'backwards'
PASS computedStyle.animationFillMode is 'backwards'
PASS style.webkitAnimationFillMode is 'backwards'
PASS computedStyle.webkitAnimationFillMode is 'backwards'
PASS style.animationFillMode is 'both'
PASS computedStyle.animationFillMode is 'both'
PASS style.webkitAnimationFillMode is 'both'
PASS computedStyle.webkitAnimationFillMode is 'both'
PASS style.animationFillMode is 'none'
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is 'none'
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is 'none, both'
PASS computedStyle.animationFillMode is 'none, both'
PASS style.webkitAnimationFillMode is 'none, both'
PASS computedStyle.webkitAnimationFillMode is 'none, both'
PASS style.animationFillMode is 'backwards, forwards'
PASS computedStyle.animationFillMode is 'backwards, forwards'
PASS style.webkitAnimationFillMode is 'backwards, forwards'
PASS computedStyle.webkitAnimationFillMode is 'backwards, forwards'
Invalid animation-fill-mode values.
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
PASS computedStyle.animationFillMode is 'none'
PASS style.webkitAnimationFillMode is ''
PASS computedStyle.webkitAnimationFillMode is 'none'
PASS style.animationFillMode is ''
Valid animation shorthand values.
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS Object.keys(style).indexOf('animation') is not -1
PASS Object.keys(style).indexOf('webkitAnimation') is -1
PASS style.animation is 'myShorthandAnim'
PASS computedStyle.animation is 'myShorthandAnim 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 0s ease 0s 1 normal none running'
PASS checkAnimationShorthandValue() is true
PASS Object.keys(style).indexOf('animation') is not -1
PASS Object.keys(style).indexOf('webkitAnimation') is -1
PASS style.animation is 'none'
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is 'none'
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is '20s none'
PASS computedStyle.animation is 'none 20s ease 0s 1 normal none running'
PASS style.webkitAnimation is '20s none'
PASS computedStyle.webkitAnimation is 'none 20s ease 0s 1 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s none'
PASS computedStyle.animation is 'myShorthandAnim 20s ease 0s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s none'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease 0s 1 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in none'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in none'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 1 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 1 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 10s'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 normal none running'
PASS computedStyle.animationDuration is '20s'
PASS computedStyle.webkitAnimationDuration is '20s'
PASS computedStyle.animationDelay is '10s'
PASS computedStyle.webkitAnimationDelay is '10s'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 10s'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 normal none running'
PASS computedStyle.animationDuration is '20s'
PASS computedStyle.webkitAnimationDuration is '20s'
PASS computedStyle.animationDelay is '10s'
PASS computedStyle.webkitAnimationDelay is '10s'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 10s paused'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 normal none paused'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s paused'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 normal none paused'
PASS computedStyle.animationDuration is '20s'
PASS computedStyle.webkitAnimationDuration is '20s'
PASS computedStyle.animationDelay is '10s'
PASS computedStyle.webkitAnimationDelay is '10s'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 5 both'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 5 normal both running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 5 both'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 5 normal both running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 5 backwards'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 5 normal backwards running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 5 backwards'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 5 normal backwards running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s 5 forwards'
PASS computedStyle.animation is 'myShorthandAnim 20s ease 0s 5 normal forwards running'
PASS style.webkitAnimation is 'myShorthandAnim 20s 5 forwards'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease 0s 5 normal forwards running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 5'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 5 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 5'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 5 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 5 reverse'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 5 reverse none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 5 reverse'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 5 reverse none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 5 reverse backwards paused'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 0s 5 reverse backwards paused'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 5 reverse backwards paused'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 0s 5 reverse backwards paused'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 normal none running, width 10s cubic-bezier(0.32, 0, 1, 1) 20s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 normal none running, width 10s cubic-bezier(0.32, 0, 1, 1) 20s 1 normal none running'
PASS computedStyle.animationDuration is '20s, 10s'
PASS computedStyle.webkitAnimationDuration is '20s, 10s'
PASS computedStyle.animationDelay is '10s, 20s'
PASS computedStyle.webkitAnimationDelay is '10s, 20s'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 10s paused, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 normal none paused, width 10s cubic-bezier(0.32, 0, 1, 1) 20s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s paused, width 10s cubic-bezier(0.32, 0, 1, 1) 20s'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 normal none paused, width 10s cubic-bezier(0.32, 0, 1, 1) 20s 1 normal none running'
PASS computedStyle.animationDuration is '20s, 10s'
PASS computedStyle.webkitAnimationDuration is '20s, 10s'
PASS computedStyle.animationDelay is '10s, 20s'
PASS computedStyle.webkitAnimationDelay is '10s, 20s'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim 20s ease-in 10s reverse paused, test 10s cubic-bezier(0.32, 0, 1, 1) 20s both'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 reverse none paused, test 10s cubic-bezier(0.32, 0, 1, 1) 20s 1 normal both running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s reverse paused, test 10s cubic-bezier(0.32, 0, 1, 1) 20s both'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 reverse none paused, test 10s cubic-bezier(0.32, 0, 1, 1) 20s 1 normal both running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'none, none'
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running, none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is 'none, none'
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running, none 0s ease 0s 1 normal none running'
PASS style.animation is 'test 20s ease-in 10s, none'
PASS computedStyle.animation is 'test 20s ease-in 10s 1 normal none running, none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is 'test 20s ease-in 10s, none'
PASS computedStyle.webkitAnimation is 'test 20s ease-in 10s 1 normal none running, none 0s ease 0s 1 normal none running'
PASS computedStyle.animationName is 'test, none'
PASS style.animation is 'none, test 20s ease-in 10s'
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running, test 20s ease-in 10s 1 normal none running'
PASS style.webkitAnimation is 'none, test 20s ease-in 10s'
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running, test 20s ease-in 10s 1 normal none running'
PASS computedStyle.animationName is 'none, test'
PASS style.animation is 'myShorthandAnim 20s ease-in 10s both paused, myShorthandAnim 20s ease-out'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 1 normal both paused, myShorthandAnim 20s ease-out 0s 1 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s both paused, myShorthandAnim 20s ease-out'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 1 normal both paused, myShorthandAnim 20s ease-out 0s 1 normal none running'
PASS style.animation is 'myShorthandAnim 20s ease-in 10s 4 backwards, myShorthandAnim2 20s ease-out 50'
PASS computedStyle.animation is 'myShorthandAnim 20s ease-in 10s 4 normal backwards running, myShorthandAnim2 20s ease-out 0s 50 normal none running'
PASS style.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 4 backwards, myShorthandAnim2 20s ease-out 50'
PASS computedStyle.webkitAnimation is 'myShorthandAnim 20s ease-in 10s 4 normal backwards running, myShorthandAnim2 20s ease-out 0s 50 normal none running'
PASS checkAnimationShorthandValue() is true
PASS style.animation is 'myShorthandAnim2 20s ease-out reverse, myShorthandAnim 20s ease-in 10s paused'
PASS computedStyle.animation is 'myShorthandAnim2 20s ease-out 0s 1 reverse none running, myShorthandAnim 20s ease-in 10s 1 normal none paused'
PASS style.webkitAnimation is 'myShorthandAnim2 20s ease-out reverse, myShorthandAnim 20s ease-in 10s paused'
PASS computedStyle.webkitAnimation is 'myShorthandAnim2 20s ease-out 0s 1 reverse none running, myShorthandAnim 20s ease-in 10s 1 normal none paused'
PASS checkAnimationShorthandValue() is true
Invalid animation shorthand values.
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is ''
PASS computedStyle.animation is 'none 0s ease 0s 1 normal none running'
PASS style.webkitAnimation is ''
PASS computedStyle.webkitAnimation is 'none 0s ease 0s 1 normal none running'
PASS style.animation is 'default 2'
PASS computedStyle.animation is 'default 0s ease 0s 2 normal none running'
PASS style.webkitAnimation is 'default 2'
PASS computedStyle.webkitAnimation is 'default 0s ease 0s 2 normal none running'
PASS successfullyParsed is true

TEST COMPLETE
